Workplace Test Code Refactoring
